requirements.txt
carla
opencv-python
numpy
pyserial